Use bigint for telegram ids and money
* Telegram ids are not guaranteed to be under 32 bits
* Money because why not. We might transact in billions someday ;)

+++ schema.sql
@@ -1,8 +1,8 @@
 CREATE TABLE transactions (
     id            bigserial,
-    payer         integer,
-    amount        integer,
-    time          integer,
+    payer         bigint,
+    amount        bigint,
+    time          bigint,
     description   text,
     group_id      bigint,
     PRIMARY KEY (id)
@@ -12,32 +12,32 @@ CREATE TABLE transactions (
 CREATE TABLE payees (
     id               bigserial,
     transaction_id   bigint references transactions(id), -- FK Transactions.Id
-    payee            integer,--# User Id
+    payee            bigint,--# User Id
     PRIMARY KEY (id)
 );
 
 CREATE TABLE authorizations (
     id               bigserial,
-    authorizer       integer, --User Id
-    authorized       integer,  --User Id
+    authorizer       bigint, --User Id
+    authorized       bigint,  --User Id
     PRIMARY KEY (id)
 );
 
 CREATE TABLE idmappings (
     username         varchar(20),
-    id               integer,
+    id               bigint,
     PRIMARY KEY (id)
 );
 
 CREATE TABLE groupmappings (
     name             text,
-    id               integer,
+    id               bigint,
     PRIMARY KEY (id)
 );
 
 CREATE TABLE belongings (
-    group_id         integer,
-    user_id          integer,
+    group_id         bigint,
+    user_id          bigint,
     PRIMARY KEY (group_id, user_id)
 );
